requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
dataclasses-json==0.5.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gptcache==0.1.21
greenlet==2.0.2
idna==3.4
install==1.3.5
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.142
langcodes==3.3.0
llama-index==0.5.27
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.2
requests==2.29.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.48
srsly==2.4.6
tenacity==8.2.2
thinc==8.1.9
tiktoken==0.3.3
tinysegmenter==0.3
tldextract==3.4.1
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.3.3
yarl==1.9.2
